Reduce Kafka consume timeout
Previously it was observed that while trying
to fetch the messages from the chopper topic
the consumer was waiting 5 minutes to timeout
and this holds up the entire polling loop so
it makes at best 1 iteration every 5 minutes.
As the loop runs quickly reduce the timeout
as the next time around we are likely to
catch it.